perf(test): split slow tsc tests from alias-bundle, ratchet aggregate budget (fixes #902)
Move validateFreeformTsc tests (bunx tsc subprocess spawning, ~14s) into
alias-bundle-tsc.spec.ts with TIMING_EXCLUSION. Ratchet AGGREGATE_TIME_BUDGET_MS
from 60s to 50s — aggregate dropped from 58.2s to 40.7s.
